requirements.txt
streamlit
numpy
tensorflow
Pillow
opencv-python-headless
streamlit-webrtc
av
